Remove CommandBase
