+++ webdev/README.md
@@ -23,24 +23,25 @@ Run a local web development server and a file system watcher that re-builds on c
 
 Usage: webdev serve [arguments] [<directory>[:<port>]]...
 -h, --help                        Print this usage information.
--r, --[no-]release                Build with release mode defaults for builders.
+    --hostname                    Specify the hostname to serve on
+                                  (defaults to "localhost")
+
+    --hot-restart                 Automatically reloads changed modules after each build and restarts your application.
+                                  Can't be used with live-reload
+
+    --live-reload                 Automatically refreshes the page after each successful build.
+                                  Can't be used with hot-restart
+
+    --log-requests                Enables logging for each request to the server.
 -o, --output                      A directory to write the result of a build to. Or a mapping from a top-level directory in the package to the directory to write a filtered build output to. For example "web:deploy".
                                   A value of "NONE" indicates that no "--output" value should be passed to `build_runner`.
                                   (defaults to "NONE")
 
--v, --verbose                     Enables verbose logging.
+-r, --[no-]release                Build with release mode defaults for builders.
     --[no-]build-web-compilers    If a dependency on `build_web_compilers` is required to run.
                                   (defaults to on)
 
-    --hostname                    Specify the hostname to serve on
-                                  (defaults to "localhost")
-
-    --log-requests                Enables logging for each request to the server.
-    --live-reload                 Automatically refreshes the page after each successful build.
-                                  Can't be used with hot-restart
-
-    --hot-restart                 Automatically reloads changed modules after each build and restarts your application.
-                                  Can't be used with live-reload
+-v, --verbose                     Enables verbose logging.
 
 Run "webdev help" to see global options.
 ```
@@ -53,16 +54,17 @@ Run builders to build a package.
 
 Usage: webdev build [arguments]
 -h, --help                        Print this usage information.
--r, --[no-]release                Build with release mode defaults for builders.
-                                  (defaults to on)
-
 -o, --output                      A directory to write the result of a build to. Or a mapping from a top-level directory in the package to the directory to write a filtered build output to. For example "web:deploy".
                                   A value of "NONE" indicates that no "--output" value should be passed to `build_runner`.
                                   (defaults to "web:build")
 
--v, --verbose                     Enables verbose logging.
+-r, --[no-]release                Build with release mode defaults for builders.
+                                  (defaults to on)
+
     --[no-]build-web-compilers    If a dependency on `build_web_compilers` is required to run.
                                   (defaults to on)
+
+-v, --verbose                     Enables verbose logging.
 
 Run "webdev help" to see global options.
 ```
